// frontend/src/app/generated/api/model/certificationViewModel.ts
/**
 * Yaam.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CertificationViewModel { 
    id: string;
    name: string;
    issuer: string | null;
    date: string;
}